refactor: condense state hook
